% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the Fatality Analysis Reporting System data for
a particular state and year on a map of that state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value assigned to a state}

\item{year}{a 4-digit year (integer or character string)}
}
\description{
This function creates a map of a given state and plots accident data
on the map by location
}
\note{
throws an error if an invalid state number is given

retruns a message if accidents for thae state and year given is 0
}
\examples{
fars_map_state(47, 2013)  #fatal accidents in TN for 2013

}
