% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Data For a List of Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of the years for which we want to get data.}
}
\description{
This function takes a vector of years, finds the corresponding data file for each year,
reads it, and returns a data frame for each year with 'MONTH' and 'year' columns.
}
\details{
The function first calls make_filename() to create a filename
for a year in the list, and then passes that filename to fars_read()
to read data from the file into a data frame. The year is added
as a column in the data frame. Finally, the MONTH and year are selected.
}
\note{
Will give a warning for any year in the list that does not have a matching
data file.
}
\examples{
fars_read_years(c('2013', '2014', '2015'))

}
