% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create Summary Data for Specified Years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of the years for which we want to get data.}

\item{return}{This function returns a summary data frame.}
}
\description{
This function takes a list of years, passes them to the
fars_read_years() function, and combines the resulting data frames
into a single summary data frame with 12 rows (one for each month)
and column for each year. Totals for each month are shown under their
respective year columns.
}
\examples{
fars_summarize_years(c('2013', '2014', '2015'))

}
