% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct a file name from a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A 4-digit year (integer or character string)}
}
\value{
This function returns a character string with the file name.
}
\description{
This function takes a 4-digit year as input and returns a filename
in the form 'accident_YEAR.csv.bz2'.
}
\examples{
make_filename(2013)

}
